+++ editor/src/main/java/org/nmox/studio/editor/lsp/LanguageServerCatalog.java
@@ -74,6 +74,12 @@ public final class LanguageServerCatalog {
                 List.of("ghcup", "install", "hls"));
         add("Zig", "zls", "install zls from github.com/zigtools/zls", List.of());
         add("Gleam", "gleam", "brew install gleam", List.of("brew", "install", "gleam"));
+        add("Nim", "nimlangserver", "nimble install nimlangserver", List.of("nimble", "install", "nimlangserver"));
+        add("D", "serve-d", "dub fetch serve-d (or brew install dub first)", List.of("dub", "fetch", "serve-d"));
+        add("Racket", "racket", "raco pkg install racket-langserver", List.of("raco", "pkg", "install", "racket-langserver"));
+        add("Elm", "elm-language-server", "npm i -g @elm-tooling/elm-language-server", List.of("npm", "i", "-g", "@elm-tooling/elm-language-server"));
+        add("ReScript", "rescript-language-server", "npm i -g @rescript/language-server", List.of("npm", "i", "-g", "@rescript/language-server"));
+        add("PureScript", "purescript-language-server", "npm i -g purescript-language-server", List.of("npm", "i", "-g", "purescript-language-server"));
         add("Erlang", "erlang_ls", "install erlang_ls from github.com/erlang-ls/erlang_ls", List.of());
         add("Clojure", "clojure-lsp", "brew install clojure-lsp/brew/clojure-lsp-native",
                 List.of("brew", "install", "clojure-lsp/brew/clojure-lsp-native"));